Landing Page 90% done
Was instructed to replicate a site created by INITS.

I setup a Next.Js project with create-next-app.

Installed bootstrap, sass, react-slick, slick carousel, and react- packages. installed a package with type definitions for react-slick so it would work with this project configured for typescript

Read the documentation for react-slick.

Got rid of the boilerplate code in the starting page route file created by Next.Js, and imported the landing page component. Also deleted the boilerplate global css styles and the svgs files.

Modified the layout file to import the bootstrap CSS and main SCSS files.

Created a react functional component to organise the different sections of the landing page

Created a navigation component to hold the nav bar. As well as modifying the layout to include the same navigation bar on every other page.

Modified the next.js.config file to allow images of a certain domain to be rendered.

Created folders and SCSS files to organise the styles for font choices, typography, buttons, icons, slider, header, footer, scss variables and the like. Imported all the complementary scss files into the main.scss file which was imported into the layout typescript-react file.

Next day:
Created a functional component for the hero section. Styled an image cover for the background.

Created eight other functional components for each of the other sections, all named according to what the section is about.

Created a News Slider component with react-slick. Fixed a code breaking bug it caused by separating the slider from the news section into It's own component and making it a client component.

Learnt how to use <svg> tags for the social media icons in a contact sections

Generated a google map embed for a specified location and displayed with an iframe tag in the footer section.

Used Next.js' Image and Link components in place of the normal html <img /> and <a> tags

Styled the each individial section in an index.scss file.

Used individual imports for react-bootstrap components to fix an incompatibility with Next.Js aswell as to stop the app from pulling the entire library to use just one of its components.

Extracted a styled button from the components that seemed to be repeated in several sections with just a change in text or border colour, for reusability.

Modified the button components to accept parameters to change the size and add bootstrap classes to the className attribute, and created a props type for the button component

Added bootstrap classes to the section containers to make them more responsive.

Created an animation for floating icons, set it up a class that uses it gave it a generic name and so it could be used by any element by adding that name to their className attribute. Created keyframes and classes for both horizontally and vertically floating icons