New App routes and Component Mods
Created an error page for unregistered routes

Created the about app route and styled it.

Migrated the sections that were components directly into the Landing Page

Created a split section component to avoid get rid of repetitive code in each section.(called DualSection)

Added iconVariants to Dual Section Component.

Added className props for the Image and Text Sections of the DualSection Component

Abstracted the types to a seperate typescript file

Created the initiatives app route

Created an InfoBox component for the  Initiatives page to avoid code duplication. I styled the infoBox from the initiative SCSS file

Adjusted the breakpoints of the News Slider component to better match the site design.

Converted a sections of lists to a slider for better responsive on smaller screens

Created container layout components to reduce code used for each section and sections with a tinted background.

Extracted the Landing Page Footer to a functional component to be reused as a layout on other pages

Created a DataBox component for the public resources page for reuseability.

Created json file to store the file names and links for the public resources page

Refactored the Button component to allow predefined variants, and links to be passed as props.

Duplicated and modified the download svg for a different style.

I styled the public resources page.